requirements.txt
aws-cdk-lib==2.110.1
constructs>=10.0.0,<11.0.0
boto3>=1.34.0
python-dotenv>=1.0.0 
aws-cdk-cli==2.1017.0
PyJWT>=2.0.0
requests>=2.31.0
pytest>=7.4.0